+++ apps/periodic/table_layout.h
@@ -7,9 +7,9 @@ namespace Periodic {
 
 class TableLayout {
 public:
-  constexpr static size_t k_numberOfColumns = 18;
-  constexpr static size_t k_numberOfRows = 9;
-  constexpr static size_t k_numberOfCells = k_numberOfColumns * k_numberOfRows;
+  constexpr static uint8_t k_numberOfColumns = 18;
+  constexpr static uint8_t k_numberOfRows = 9;
+  constexpr static uint8_t k_numberOfCells = k_numberOfColumns * k_numberOfRows;
 
   enum class Direction : uint8_t {
     IncreasingZ,
@@ -19,10 +19,10 @@ public:
   };
 
   static AtomicNumber NextElement(AtomicNumber z, Direction direction);
-  static void PositionForElement(AtomicNumber z, size_t * column, size_t * row);
-  static void PositionForCell(size_t cell, size_t * column, size_t * row);
-  static AtomicNumber ElementInCell(size_t cellIndex);
-  static size_t CellForElement(AtomicNumber z);
+  static void PositionForElement(AtomicNumber z, uint8_t * column, uint8_t * row);
+  static void PositionForCell(uint8_t cell, uint8_t * column, uint8_t * row);
+  static AtomicNumber ElementInCell(uint8_t cellIndex);
+  static uint8_t CellForElement(AtomicNumber z);
 };
 
 }
